fix: moves config types into graphql-armor-types package
