+++ packages/graphql-armor/src/index.ts
@@ -1,4 +1,2 @@
-export type { GraphQLArmorConfig } from './config';
-
 export { ApolloArmor } from './apollo/armor';
 export { EnvelopArmor, EnvelopArmorPlugin } from './envelop/armor';
